requirements.txt
youtube-transcript-api==1.2.2
pytube>=15.0.0
rapidfuzz>=3.0.0
pydantic>=2.0.0
openai>=1.0.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
bert-score>=0.3.13
rouge-score>=0.1.2
streamlit>=1.36.0